move non_domain_sort code to operators/selection
