+++ src/evox/operators/__init__.py
@@ -1,2 +1,5 @@
-from .non_dominated_sort import non_dominated_sort
-from .crowding_distance_sort import crowding_distance, crowding_distance_sort
+from .selection.non_dominate import (
+    non_dominated_sort,
+    crowding_distance,
+    crowding_distance_sort,
+)
